Add multipart support
- Cleaning (Intellij code analysis)

+++ view/src/main/java/com/jsql/view/swing/panel/preferences/SpinnerMouseWheelListener.java
@@ -14,7 +14,7 @@ public class SpinnerMouseWheelListener implements MouseWheelListener {
         JSpinner source = (JSpinner) e.getComponent();
         SpinnerNumberModel model = (SpinnerNumberModel) source.getModel();
         Integer oldValue = (Integer) source.getValue();
-        var intValue = oldValue.intValue() - e.getWheelRotation() * model.getStepSize().intValue();
+        var intValue = oldValue - e.getWheelRotation() * model.getStepSize().intValue();
         int max = (Integer) model.getMaximum(); //1000
         int min = (Integer) model.getMinimum(); //0
         if (min <= intValue && intValue <= max) {
